% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentree_taxonomy.R
\name{check_ott_input}
\alias{check_ott_input}
\title{Check input for other functions}
\usage{
check_ott_input(input = NULL, ott_ids = NULL)
}
\arguments{
\item{input}{Optional. A character vector of names.}

\item{ott_ids}{If not NULL, it takes this argument and ignores input. A
numeric vector of ott ids obtained with \code{\link[rotl:taxonomy_taxon_info]{rotl::taxonomy_taxon_info()}} or
\code{\link[rotl:tnrs_match_names]{rotl::tnrs_match_names()}} or \code{\link[=tnrs_match]{tnrs_match()}}.}
}
\value{
A named numeric vector of valid Open Tree Taxonomy (OTT) ids.
}
\description{
\verb{check_ott_input is currently used in functions }\code{\link[=get_ott_clade]{get_ott_clade()}},
\code{\link[=get_ott_children]{get_ott_children()}}.
}
\details{
By default, it uses the \code{ott_id} argument if it is not NULL.
}
