% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_phylo}
\alias{is_phylo}
\title{Checks if \code{phy} is a \code{phylo} object and/or a chronogram.}
\usage{
is_phylo(phy = NULL, brlen = FALSE, dated = FALSE)
}
\arguments{
\item{phy}{A \code{phylo} object.}

\item{brlen}{Boolean. If \code{TRUE} it checks if \code{phylo} object has branch lengths.}

\item{dated}{Boolean. If \code{TRUE} it checks if \code{phylo} object is ultrametric.}
}
\value{
Nothing
}
\description{
Checks if \code{phy} is a \code{phylo} object and/or a chronogram.
}
