% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{phylo_tiplabel_space_to_underscore}
\alias{phylo_tiplabel_space_to_underscore}
\title{Convert spaces to underscores in trees.}
\usage{
phylo_tiplabel_space_to_underscore(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A \code{phylo} object.
}
\description{
Convert spaces to underscores in trees.
}
