% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{phylo_tiplabel_underscore_to_space}
\alias{phylo_tiplabel_underscore_to_space}
\title{Convert underscores to spaces in trees.}
\usage{
phylo_tiplabel_underscore_to_space(phy)
}
\arguments{
\item{phy}{A \code{phylo} object.}
}
\value{
A \code{phylo} object.
}
\description{
Convert underscores to spaces in trees.
}
